working centralized and distributed fl
